Instala una acción GitHub que genera tu sitio en la rama gh-pages

Instala una acción GitHub que crea el sitio pkgdown de tu paquete en la rama gh-pages de tu repositorio.

Uso

ros_actions_pkgdown_branch(pkg = ".", overwrite = TRUE)

Argumentos

pkg

Ruta a la raíz del paquete.

overwrite

Sobreescribe la acción si ya se encuntra instalada.

Ver también

Sitio creado con pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.