Build your pkgdown website locally

A wrapper of pkgdown::build_site()

Usage

ros_build(pkg = ".", ...)

Arguments

pkg

Path to package.

...

Arguments passed on to pkgdown::build_site

examples

Run examples?

run_dont_run

Run examples that are surrounded in \dontrun?

seed

Seed used to initialize so that random examples are reproducible.

lazy

If TRUE, will only rebuild articles and reference pages if the source is newer than the destination.

override

An optional named list used to temporarily override values in _pkgdown.yml

preview

If TRUE, or is.na(preview) && interactive(), will preview freshly generated section in browser.

devel

Use development or deployment process?

If TRUE, uses lighter-weight process suitable for rapid iteration; it will run examples and vignettes in the current process, and will load code with pkgload::load_all().

If FALSE, will first install the package to a temporary library, and will run all examples and vignettes in a new process.

build_site() defaults to devel = FALSE so that you get high fidelity outputs when you building the complete site; build_reference(), build_home() and friends default to devel = TRUE so that you can rapidly iterate during development.

new_process

If TRUE, will run build_site() in a separate process. This enhances reproducibility by ensuring nothing that you have loaded in the current process affects the build process.

install

If TRUE, will install the package in a temporary library so it is available for vignettes.

document

Deprecated Use devel instead.

Details

This function would ask for confirmation to check your _pkgdown.yml file and would create your pkgdown site on the docs/ folder.

Unlike ros_actions_pkgdown_branch(), the user would trigger the build and it is possible to control it via the parameters of pkgdown::build_site().

See also

Site built with pkgdown 2.0.1.

Template rostemplate by dieghernan, based on Bootstrapious.