Build your pkgdown website locally

A wrapper of pkgdown::build_site

ros_build(pkg = ".", ...)

Arguments

pkg, ...
See ?pkgdown::build_site.

Details

This function would ask for confirmation to check your _pkgdown.yml file and would create your pkgdown site on the /docs folder.

Unlike ros_gh_actions_branch, the user would trigger the build and it is possible to control it via the parameters of pkgdown::build_site.

See also

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.